add --log_format_version
the version 1 is compatible with logstash event format 1.

additionally, put everything into a top field, "sensu".
this avoids field name collision. when it happens, not only it makes
things complicated but also creates a race condition, which is hard to
notice. if other applications use "level" as an int field and win the
race, elasticsearch refuse to accept the request if the field is not
mapped as string.